libghostty: move selection functions to selection doxygen group
