+++ src/terminal/c/main.zig
@@ -170,10 +170,10 @@ pub const terminal_mode_get = terminal.mode_get;
 pub const terminal_mode_set = terminal.mode_set;
 pub const terminal_get = terminal.get;
 pub const terminal_get_multi = terminal.get_multi;
-pub const terminal_selection_adjust = terminal.selection_adjust;
-pub const terminal_selection_order = terminal.selection_order;
-pub const terminal_selection_ordered = terminal.selection_ordered;
-pub const terminal_selection_contains = terminal.selection_contains;
+pub const terminal_selection_adjust = selection.adjust;
+pub const terminal_selection_order = selection.order;
+pub const terminal_selection_ordered = selection.ordered;
+pub const terminal_selection_contains = selection.contains;
 pub const terminal_grid_ref = terminal.grid_ref;
 pub const terminal_grid_ref_track = terminal.grid_ref_track;
 pub const terminal_point_from_grid_ref = terminal.point_from_grid_ref;
